+++ src/media/visual/mod.rs
@@ -6,7 +6,6 @@
 #![doc = crate::_doc!(hr)]
 //
 
-#[cfg(feature = "color")]
 pub mod color;
 #[cfg(feature = "draw")]
 pub mod draw;
@@ -15,7 +14,7 @@ pub mod image;
 
 crate::structural_mods! { // _pub_mods, _crate_internals
     _pub_mods {
-        #[cfg(feature = "color")] pub use super::color::_all::*;
+        pub use super::color::_all::*;
         #[cfg(feature = "draw")] pub use super::draw::_all::*;
         pub use super::image::_all::*;
         // pub use super::video::_all::*;