+++ src/run/regime/cap/mod.rs
@@ -3,7 +3,7 @@
 //! Runtime capabilities.
 //
 
-mod definitions; // RunCap[Audio|Image|Input|System|Window]
+mod definitions; // RunCap[Audio|Color|Image|Input|System|Window]
 
 crate::structural_mods! { // _mods
     _mods {